chore: restructure package layout
