+++ pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
 ]
 
 [project.scripts]
-arxiv-recommend = "arxiv_recommender.cli:main"
+arxiv-recommend = "arxiv_recommender.interfaces.cli:main"
 
 [project.optional-dependencies]
 dev = ["pytest>=7.0", "ruff>=0.1", "mypy>=1.0", "pre-commit>=3.0", "types-requests"]
