+++ src/arxiv_recommender/interfaces/__init__.py
@@ -1,0 +1,1 @@
+"""User-facing interfaces for the arXiv recommender."""
