+++ src/arxiv_recommender/schemas/config.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, Field, field_validator
 
-from arxiv_recommender.embedding_config import (
+from arxiv_recommender.text_vectorization.config import (
     AutoSetting,
     NormalizeEmbeddingsSetting,
     PoolingStrategy,
