+++ src/arxiv_recommender/text_vectorization/pooling.py
@@ -2,7 +2,7 @@ from collections.abc import Callable
 
 import torch
 
-from arxiv_recommender.embedding_config import PoolingStrategy
+from arxiv_recommender.text_vectorization.config import PoolingStrategy
 
 Pooler = Callable[[torch.Tensor, torch.Tensor], torch.Tensor]
